passou em todos os casos de teste

+++ trabalho01/t1.c
@@ -2,8 +2,8 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-void completa_matriz(int **m, int n){
-	int l, c;
+void completa_matriz(unsigned long int **m, int n){
+	unsigned long int l, c;
 	for (l = 0; l < n; l++){
 		for (c = 0; c <= l; c++){
 			if (c == 0 || l == c){
@@ -16,11 +16,11 @@ void completa_matriz(int **m, int n){
 	}
 }
 
-void imprime_matriz(int **m, int n){
-	int i, j;
+void imprime_matriz(unsigned long int **m, int n){
+	unsigned long int i, j;
 	for (i = 0; i < n; i++){
 		for (j = 0; j <= i; j++){
-			printf("%d\n", m[i][j]);
+			printf("%lu\n", m[i][j]);
 		}
 	}
 }
@@ -29,9 +29,9 @@ int main(void){
 	int n, i;
 	scanf(" %d", &n);
 	
-	int **m = (int **) malloc((n+1)*sizeof(int *));
+	unsigned long int **m = (unsigned long int **) malloc((n+1)*sizeof(unsigned long int *));
 	for (i = 0; i < n; i++){
-		m[i] = (int *) malloc((i+1)*sizeof(int));
+		m[i] = (unsigned long int *) malloc((i+1)*sizeof(unsigned long int));
 	}
 	
 	completa_matriz(m, n);
